// src/app/modules/comprobante/registro-comprobante/model/datos-factura-compra.model.ts
export interface DatosFacturaCompra {
    id?: number;
    sucursalId?: number;
    tipoFormulario?: string;
    numeroComprobante?: number;
    tipoCompra?: string;
    fechaFactura?: string;
    nit?: string;
    razonSocial?: string;
    eTicket?: number;
    numeroFactura?: string;
    numeroDui?: number;
    codigoAutorizacion?: string;
    fechaComprobante?: string;
    importeTotalCompra?: number;
    importeSinCf?: number;
    subtotal?: number;
    creditoFiscal?: number;
    codigoControl?: string;
    importeNeto?: number;
    totalIva?: number;
    importeComprasGravadasTasaCero?: number;
    importeIcei?: number;
    importeIehd?: number;
    importeIpj?: number;
    tasa?: number;
    otrosNscf?: number;
    importeExentos?: number;
    descuentoBonificacionRebajaIva?: number;
    importeGiftCard?: number;
    importeBaseCf?: number;
    sinDerCredFiscal?: boolean;
    gasolina?: boolean;
    fechaOperacion?: string;
}